Consolidate Dashboard removal flows into one file
Fold removeDashboardUpgradeFlow into cancel-dashboard-purchase-flow.ts,
matching the legacy cancel-purchase-flow.ts layout. Share the
upgrades/<id>/cancel response wait via waitForUpgradeCancel.
