+++ packages/calypso-e2e/src/lib/flows/index.ts
@@ -7,4 +7,3 @@ export * from './lohp-theme-signup-flow';
 // Newer style flows (without a class)
 export * from './cancel-purchase-flow';
 export * from './cancel-dashboard-purchase-flow';
-export * from './remove-dashboard-upgrade-flow';
